requirements.txt
Django
djangorestframework
psycopg2-binary
django-cors-headers 